separeted the dom from the logic and fix restart match
