+++ index.html
@@ -41,6 +41,6 @@
       <p class="move"></p>
     </div>
 
-    <script src="main.js"></script>
+    <script type="module" src="main.js"></script>
   </body>
 </html>
